[language] Finishing refactor of language code
