+++ test/test_language.py
@@ -1,14 +1,17 @@
 import threading
 import unittest
+
+from anytree import RenderTree
+
 from pycram.designators.action_designator import *
 from pycram.enums import ObjectType, State
 from pycram.plan_failures import PlanFailure
 from pycram.pose import Pose
-from pycram.new_language import Sequential, Language, Parallel, TryAll, TryInOrder
+from pycram.language import Sequential, Language, Parallel, TryAll, TryInOrder
 from pycram.process_module import simulated_robot
 import test_bullet_world
 
-from pycram.task import Code
+from pycram.language import Code
 
 
 class LanguageTestCase(test_bullet_world.BulletWorldTest):
@@ -92,7 +95,6 @@ class LanguageTestCase(test_bullet_world.BulletWorldTest):
         act2 = PickUpAction(BelieveObject(names=["milk"]), ["left"], ["front"])
 
         self.assertRaises(TypeError, lambda: act ^ act2)
-
 
     def test_perform_desig(self):
         act = NavigateAction([Pose([1, 1, 0])])
